examples/langchain_tools/requirements.txt
tourist
langchain_aws
langgraph
python-dotenv
beautifulsoup4